refactor: extract sync lock management from SyncService to SyncLock class
Move acquireSyncLock and releaseSyncLock logic into a new dedicated
SyncLock class (vscode-extension/src/backend/services/syncLock.ts).

- SyncLock exposes acquire(backend?, serverUrl?) and release(backend?)
- SyncLock.STALE_MS holds the stale-threshold constant (was SYNC_LOCK_STALE_MS)
- SyncService constructs a SyncLock instance and delegates both methods
- No visible behaviour change; all existing tests continue to pass
